Remove all TokenTransact usages
