+++ hedera-node/src/main/java/com/hedera/services/fees/calculation/crypto/txns/CryptoTransferResourceUsage.java
@@ -24,11 +24,9 @@ import com.hedera.services.context.primitives.StateView;
 import com.hedera.services.fees.calculation.TxnResourceUsageEstimator;
 import com.hedera.services.usage.SigUsage;
 import com.hedera.services.usage.crypto.CryptoTransferUsage;
-import com.hedera.services.usage.token.TokenTransactUsage;
 import com.hederahashgraph.api.proto.java.FeeData;
 import com.hederahashgraph.api.proto.java.TransactionBody;
 import com.hederahashgraph.exception.InvalidTxBodyException;
-import com.hederahashgraph.fee.CryptoFeeBuilder;
 import com.hederahashgraph.fee.SigValueObj;
 
 import java.util.function.BiFunction;
